chinese_proverbs/requirements.txt
alembic==1.7.5
aniso8601==9.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
binary==1.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
JsonForm==0.0.2
jsonschema==4.4.0
JsonSir==0.0.2
Mako==1.1.6
MarkupSafe==2.0.1
migrate==0.3.8
numpy==1.21.5
pandas==1.3.5
psycopg2-binary==2.9.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
Python-EasyConfig==0.1.7
pytz==2021.3
PyYAML==6.0
requests==2.27.1
Resource==0.2.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
